Turns out cellsize is a double
git-svn-id: file:///home/rick/.svn/richdem@175 c4c8cafc-911d-468f-9f86-3f4f8cd09a06

Former-commit-id: a997594546e6ba20b5ca7f8eb44e85ca85bd406a

+++ data_structures.h
@@ -12,7 +12,7 @@
 template <class T>
 class array2d : public boost::numeric::ublas::matrix<T>{
 	public:
-		int cellsize;
+		double cellsize;
 //		std::string xllcorner,yllcorner;	//TODO: Should be string
 		double xllcorner,yllcorner;
 		long data_cells;
